re-arranging theme files to make it easy to add/modify styles
